AmazonSP: clean old code & tests
